Remove iio_context parameter from Context::getAllDevices and use m_context instead
The iio_context is already available for the ContextImpl object
in it's private m_context member. There is no need to pass
another iio_context to this method. Moreover, the iio_context is
not available to the user if the Context was obtained with
deviceOpen(const char*) or deviceOpen(). The only solution would
be for the user to obtain another iio_context pointer, but that
is also not possible while the current one is open.

Signed-off-by: Mihai Olteanu <Mihai.Olteanu@analog.com>