Add Lidar
Add Lidar to the list of boards discoverable by the ContextBuilder.

Use the DeviceGeneric and DeviceIn classes to implement the Lidar
requirements from the Lidar API page [1] and expose the relevant
interfaces to the user.

Initialize the board in the same way the iio-oscilloscope does with
the Lidar plugin (some code is actually identical with that), but also
enable Channel4 by default, which should always be enabled.

Reading the buffer returns a list of all the enabled channels together
with the requested number of samples.

Add lidar example in the examples folder

[1] Lidar API

Signed-off-by: Mihai.Olteanu Mihai.Olteanu@analog.com