Device Generic, In, Out: Add changes to the channel scanning system.
Previously, DeviceGeneric(the class which handles the iio_device
interaction) scanned channels using their index (for DMMs) and using
their name/ID for all the other devices (ADCs, DACs). It assumed
that all the ADC or DAC channels are called "voltageX" which is not right.
In order to correctly use the channels, it scans them all based on the
index. But this is not enough, the list of channels is separated in input
and output channels and sorted based on the ID of each channel (could
be improved) because the order in iio_device might change.

Signed-off-by: Alexandra.Trifan <Alexandra.Trifan@analog.com>