M2kAnalogIn: Enable all the channels before refilling a buffer.
The "m2k-adc" requires all the channels to be enabled before
acquiring data. All the channels must be enabled before a
"iio_buffer_refill" call and then disabled if necessary
(brought to their previous state).
This should not be in the Buffer implementation, since other
boards might not have the same behaviour.

Signed-off-by: Alexandra.Trifan <Alexandra.Trifan@analog.com>