M2kHardwareTriggerImpl: Fix MSVC undefined behaviour for methods throwing instead of returning.
These methods in particular only throw an exception to the user, letting
him know that the operation is not possible. When building with CMake-GUI
and Microsoft Visual Studio on Windows, the compiler complains about the
missing return value. To ensure that these throw we call "throw" directly,
without using our own "throw" wrapper.

Signed-off-by: Alexandra.Trifan <Alexandra.Trifan@analog.com>