Buffer & channel - general restructure: Drop pointers to implementation for the iio_channel and iio_buffer wrappers.
These classes should not be exposed by the library. Their only use
is to interface between the iio_channel/iio_buffer and the tools based on
them. They provide basic functionality for all the libiio features.
We can drop the Pimpl idea for these classes if we don't expose them.
Also, these classes need to be moved outside of the official "include" dir.

Signed-off-by: Alexandra Trifan <Alexandra.Trifan@analog.com>