M2kHardwareTrigger: Fix digital source confusion when retrieving its value.
The digital trigger source can be either "trigger-logic" or "trigger-in"
(if we're talking about IIO attributes). In libm2k, we added
SRC_TRIGGER_IN (trigger-logic), SRC_ANALOG_IN (trigger-in) and SRC_NONE.
The third should also point to "trigger-logic", because that is the default
value for this attribute. However, we didn't check this when calling
"getDigitalSource". This commit should verify whether there is a trigger
condition attached to this source and return SRC_NONE or SRC_TRIGGER_IN
accordingly.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>