AnalogIn & Out: Replace getAvailableSamples method.
Remove for utils the method for getting the available samples.
Change the implementation of getAvailableSamples from AnalogIn & Out
such that the new method getAvailableAttributeValues will be used.
The method getAvailableAttributeValues returns a vector of strings
for any type of attribute. A conversion to double is performed for
the values of the sampling frequency.

Signed-off-by: Teo Perisanu <Teo.Perisanu@analog.com>