Revert "Destroy the buffer after canceling any operation."
This reverts commit 941df777ab609c2b7d09517079b7e02aa415b7aa.

The buffer canceling leaves the buffer in an invalid state. It should be
used only to disrupt a blocked operation.  It should not be used if the
buffer is not currently performing a blocking operation.

However, this operation should not include buffer destroying. The
destroyal part is done by managing the buffer state after refill/push
operations.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>