bindings/CMakeLists.txt: Use the OLD cmake_policy for CMP0078.
This policy is related to the SWIG generated library name. For older
versions, the name of the module and the target would be modify inside the
swig_add_library function.
The new version would keep the exact same name for the generated target.
A constant fix for both versions would mean a change of name for the python
module (which is not accepted right now).

Until we find a better solution we will use the OLD behaviour for all the
CMake versions (since this was proven to work until now) and we'll work
on finding a better solution until the moment of deprecation.

Signed-off-by: AlexandraTrifan <Alexandra.Trifan@analog.com>