ContextBuilder: Open the available context with m2kOpen().
In later versions of libm2k, calling m2kOpen() on Linux distributions could
result in an exception, even though an USB interface would be available.
This problem was caused by trying to establish a connection to the first
USB interface.
In this commit, the verification is made by using iio_create_context, and
checking if the context is null. Also stderr was suppressed inside this
function.

Signed-off-by: Teo Perisanu <Teo.Perisanu@analog.com>