libm2k.iss: Use the "replacesameversion" flag for .exe and .dlls.
From the official documentation: "When this flag is used and the file
already exists on the user's system and it has the same version number
as the file being installed, Setup will compare the files and replace
the existing file if their contents differ.
The default behavior (i.e. when this flag isn't used) is to not replace
an existing file with the same version number."

Signed-off-by: Alexandra Trifan <Alexandra.Trifan@analog.com>