M2kImpl: block the digital/analog interface until started
For the case where there is a trigger either on the analog interface or digital interface, we need to disable the triggering interface in order to block the acquisition. When both analog and digital interfaces are started we then release the corresponding trigger. Otherwise there will be a larger delay due to the analog acquisition being started much sooner than the digital acquisition.

Signed-off-by: DanielGuramulta <Daniel.Guramulta@analog.com>