libm2k.iss.cmakein: Copy the C++ -> C# wrapper DLL into System32 when installing.
The libm2k-sharp.dll is added as a reference to a C# project, so the
compiler will manually copy it to the build directory.
But the libm2k-sharp-cxx-wrap.dll needs to be copied manually in the
build directory, and can't be added as a reference to a C# project.
This raises lots of issues. If we place it in the system folder, any DLL
that needs it, will find it in the Path.

Signed-off-by: Alexandra Trifan <Alexandra.Trifan@analog.com>