debug: simulate new FW - calibration in SW
- in new firmware version the calibration is done in SW. Params will be
written in the device attr but not used at the HW level. Raw samples are
sent without calibration and the calibration is done in SW.
- this patch is for debug purpose only. It updates calibration params
only for the class instance but not in the device attr. This way we simulate
calibration where samples are sent without calibration at HW level.

Signed-off-by: Adrian Stanea <Adrian.Stanea@analog.com>